Move split integration tests to 3
