+++ cypress/integration/group1/disabledAccountControls.ts
@@ -15,7 +15,7 @@
  *     See the License for the specific language governing permissions and
  *     limitations under the License.
  */
-import { resetDB, setTokenUserViewPort } from '../support/commands';
+import { resetDB, setTokenUserViewPort } from '../../support/commands';
 
 describe('Go to disabled Dockstore Account Controls', function () {
   resetDB();
