+++ cypress/integration/group1/dropdown.ts
@@ -15,7 +15,7 @@
  *    See the License for the specific language governing permissions and
  *    limitations under the License.
  */
-import { setTokenUserViewPort } from '../support/commands';
+import { setTokenUserViewPort } from '../../support/commands';
 
 describe('Dropdown test', function () {
   // TODO: GitLab tests are commented out
