+++ cypress/integration/group2/VersionModal.ts
@@ -13,7 +13,7 @@
  *    See the License for the specific language governing permissions and
  *    limitations under the License.
  */
-import { setTokenUserViewPort, resetDB } from '../support/commands';
+import { setTokenUserViewPort, resetDB } from '../../support/commands';
 
 describe('Public Version Modal', function () {
   resetDB();
