+++ cypress/integration/group2/toolDetails.ts
