+++ cypress/integration/group2/workflowDetails.ts
