+++ cypress/integration/group3/sharedWorkflows.ts
@@ -13,7 +13,7 @@
  *    See the License for the specific language governing permissions and
  *    limitations under the License.
  */
-import { goToTab, goToUnexpandedSidebarEntry, resetDB, setTokenUserViewPort } from '../support/commands';
+import { goToTab, goToUnexpandedSidebarEntry, resetDB, setTokenUserViewPort } from '../../support/commands';
 
 describe('Shared with me workflow test from my-workflows', function () {
   resetDB();
