+++ cypress/integration/group3/starEntry.ts
@@ -13,7 +13,7 @@
  *     See the License for the specific language governing permissions and
  *     limitations under the License.
  */
-import { resetDB, setTokenUserViewPort } from '../support/commands';
+import { resetDB, setTokenUserViewPort } from '../../support/commands';
 
 describe('Tool and Workflow starring', function () {
   resetDB();
